Add option for query author to delete query result
After each query is processed and responded to, mallard sticks a
:wastebasket: reaction at the bottom of the result message. If the query
author adds that reaction, the result message is deleted. Author-query
relationships are stored in a list with a capacity of 50. (As such, if
mallard goes down, the information is lost, however this is acceptable
as the information is non-imperative.)